+++ sklearn/preprocessing/__init__.py
@@ -29,7 +29,7 @@ from .label import LabelBinarizer
 from .label import LabelEncoder
 from .label import MultiLabelBinarizer
 
-from .target import TransformTargetRegressor
+from .target import TransformedTargetRegressor
 from .imputation import Imputer
 
 
@@ -48,7 +48,7 @@ __all__ = [
     'OneHotEncoder',
     'RobustScaler',
     'StandardScaler',
-    'TransformTargetRegressor',
+    'TransformedTargetRegressor',
     'add_dummy_feature',
     'PolynomialFeatures',
     'binarize',